feat: AIモデルの精度向上（学習方法の改善とAPI修正）
