+++ backend/race_prediction_model.pkl
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d4aec9e51ac6575d62ef14e0cf2100156d1a10d72d52218a809e7a3d302b68f1
-size 143648185
+oid sha256:1bda57eeaa70c60e1be58f63033aab085b2667a658bc24cfa6b45cfe1d3fdd3e
+size 143567385
